+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Forward property-targetable attributes from eligible fields to generated properties
+- Opt into child `INotifyPropertyChanged` notifications with `NotifyAlso.NotifyOnSubPropertyChanged`
+
 ## [1.7.0] - 2026-08-08
 
 ### Added